+++ Changelog.md
@@ -2,6 +2,8 @@
 
 ### Version 1.16.3 - 1.1.8
  - added raccoon entity
+     - they destroy the ostrich egg in a nest
+     - they destroy turtle eggs
 
 ### Version 1.16.3 - 1.1.7
  - added mantaray entity